Enable Chromium to run on devcontainer
The Chromium browser has to run in headless mode when we are running in
a devcontainer (e.g. GitHub Codespaces).

Also, we need to run the container without [seccomp][1] restrictions,
otherwise we receive an error when running the specifications, saying
that there is no usable sandbox for Chromium to run in.

It would be nice if we could use a [custom seccomp profile][2], rather
than removing all seccomp restrictions entirely, but it appears that
Codespaces doesn't currently allow custom seccomp profiles. Certainly
the [docs for the (now-obsolete) Visual Studio Codespaces say that][3].

[1]: https://docs.docker.com/engine/security/seccomp/
[2]: https://github.com/jlund/docker-chrome-pulseaudio/issues/8#issue-166464652
[3]: https://docs.microsoft.com/en-us/visualstudio/codespaces/reference/configuring#docker-properties