Set up default Gauge Taiko project (#9)
The Gauge specifications pass successfully on the GitHub Codespaces
environment.

* Add default Gauge and Taiko project files

These files were all generated by running: `gauge init js`

* Fix typo in comment

* Ignore generated Taiko Dockerfile linting issues

We are not currently using this Dockerfile, so we don't need to fix the
issues reported in the linting at this point of time.  If we do ever use
this Dockerfile we should then unignore the issues and address the
issues reported.

* Enable Chromium to run on devcontainer

The Chromium browser has to run in headless mode when we are running in
a devcontainer (e.g. GitHub Codespaces).

Also, we need to run the container without [seccomp][1] restrictions,
otherwise we receive an error when running the specifications, saying
that there is no usable sandbox for Chromium to run in.

It would be nice if we could use a [custom seccomp profile][2], rather
than removing all seccomp restrictions entirely, but it appears that
Codespaces doesn't currently allow custom seccomp profiles. Certainly
the [docs for the (now-obsolete) Visual Studio Codespaces say that][3].

[1]: https://docs.docker.com/engine/security/seccomp/
[2]: https://github.com/jlund/docker-chrome-pulseaudio/issues/8#issue-166464652
[3]: https://docs.microsoft.com/en-us/visualstudio/codespaces/reference/configuring#docker-properties

* Add Gauge file associations to VSCode

* Install Gauge VSCode extension on devcontainer

* Install npm packages on devcontainer